+++ age/CHANGELOG.md
@@ -14,6 +14,7 @@ to 1.0.0 are beta releases.
 
 ### Changed
 - MSRV is now 1.65.0.
+- Migrated to `base64 0.21`.
 
 ## [0.9.2] - 2023-06-12
 ### Added